Update Vite configuration and add new dependencies
This commit updates the configuration for Vue and Laravel in vite.config.js, and adds new dependencies for Vue and TailwindCSS. The docker settings for Vite have also been configured. Inertia.js and Laravel Breeze have been added as dependencies in package.json, along with the necessary Vue and Tailwind configuration for Laravel Breeze.